move tile_peptides to util file
